+++ common/src/main/java/com/github/yajatkaul/mega_showdown/MegaShowdown.java
@@ -2,40 +2,31 @@ package com.github.yajatkaul.mega_showdown;
 
 import com.github.yajatkaul.mega_showdown.block.MegaShowdownBlockEntities;
 import com.github.yajatkaul.mega_showdown.block.MegaShowdownBlocks;
+import com.github.yajatkaul.mega_showdown.command.MegaShowdownCommands;
 import com.github.yajatkaul.mega_showdown.components.MegaShowdownDataComponents;
 import com.github.yajatkaul.mega_showdown.config.MegaShowdownConfig;
 import com.github.yajatkaul.mega_showdown.creative.MegaShowdownTabs;
-import com.github.yajatkaul.mega_showdown.datapack.MegaShowdownDatapackRegister;
 import com.github.yajatkaul.mega_showdown.event.EventRegister;
 import com.github.yajatkaul.mega_showdown.item.MegaShowdownItems;
 import com.github.yajatkaul.mega_showdown.networking.server.MegaShowdownNetworkHandlerServer;
 import com.github.yajatkaul.mega_showdown.screen.MegaShowdownMenuTypes;
 import com.github.yajatkaul.mega_showdown.sound.MegaShowdownSounds;
-import com.github.yajatkaul.mega_showdown.status.MegaShowdownStatusEffects;
-import com.github.yajatkaul.mega_showdown.utils.DelayedTicker;
-import com.github.yajatkaul.mega_showdown.utils.ShowdownItemsLoad;
-import dev.architectury.event.events.common.LifecycleEvent;
-import dev.architectury.event.events.common.TickEvent;
-import net.minecraft.resources.ResourceLocation;
+import dev.architectury.event.events.common.CommandRegistrationEvent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public final class MegaShowdown {
     public static final String MOD_ID = "mega_showdown";
-    public static final Logger LOGGER = LoggerFactory.getLogger("MegaShowdown");
-
-    public static ResourceLocation msdResource(String id) {
-        return ResourceLocation.fromNamespaceAndPath(MegaShowdown.MOD_ID, id);
-    }
+    public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
 
     public static void init() {
         MegaShowdownConfig.load();
-        MegaShowdownStatusEffects.register();
         MegaShowdownDataComponents.register();
         MegaShowdownBlocks.register();
         MegaShowdownBlockEntities.register();
         MegaShowdownItems.register();
         MegaShowdownSounds.register();
+        CommandRegistrationEvent.EVENT.register(MegaShowdownCommands::registerCommands);
 
         MegaShowdownTabs.register();
 
@@ -44,12 +35,5 @@ public final class MegaShowdown {
         MegaShowdownMenuTypes.register();
 
         EventRegister.register();
-
-        LifecycleEvent.SERVER_STARTING.register((minecraftServer) -> {
-            MegaShowdownDatapackRegister.registerShowdownDatapackItems(minecraftServer);
-            ShowdownItemsLoad.load();
-        });
-
-        TickEvent.SERVER_PRE.register((server) -> DelayedTicker.runAll());
     }
 }