+++ common/src/main/java/com/github/yajatkaul/mega_showdown/client/ClientEventRegister.java
@@ -1,0 +1,8 @@
+package com.github.yajatkaul.mega_showdown.client;
+
+public class ClientEventRegister {
+    public static void register() {
+        CobbleClientEvents.register();
+        ClientMinecraftRegister.register();
+    }
+}