+++ common/src/main/java/com/github/yajatkaul/mega_showdown/event/EventRegister.java
@@ -4,5 +4,6 @@ public class EventRegister {
     public static void register() {
         CobbleEvents.register();
         AccessoriesEvent.register();
+        MinecraftEvents.register();
     }
 }