+++ fabric/src/main/java/com/github/yajatkaul/mega_showdown/fabric/MegaShowdownFabric.java
@@ -4,7 +4,6 @@ import com.cobblemon.mod.common.Cobblemon;
 import com.cobblemon.mod.common.battles.runner.graal.GraalShowdownService;
 import com.cobblemon.mod.relocations.graalvm.polyglot.Value;
 import com.github.yajatkaul.mega_showdown.MegaShowdown;
-import com.github.yajatkaul.mega_showdown.command.MegaShowdownCommands;
 import com.github.yajatkaul.mega_showdown.datapack.MegaShowdownDatapackRegister;
 import com.github.yajatkaul.mega_showdown.fabric.datapack.DatapackRegistry;
 import com.github.yajatkaul.mega_showdown.gimmick.MaxGimmick;
@@ -12,7 +11,6 @@ import kotlin.Unit;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.biome.v1.BiomeModifications;
 import net.fabricmc.fabric.api.biome.v1.BiomeSelectors;
-import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceKey;
@@ -34,18 +32,14 @@ public final class MegaShowdownFabric implements ModInitializer {
         DatapackRegistry.register();
         generateModWorldGen();
 
-        CommandRegistrationCallback.EVENT.register(MegaShowdownCommands::registerCommands);
-
-        ServerLifecycleEvents.SERVER_STARTED.register((server) -> {
-            Cobblemon.INSTANCE.getShowdownThread().queue(showdownService -> {
-                if (showdownService instanceof GraalShowdownService service) {
-                    Value receiveMoveDataFn = service.context.getBindings("js").getMember("receiveCustomGmaxMove");
-                    for (MaxGimmick gmax : MegaShowdownDatapackRegister.GMAX_REGISTRY) {
-                        receiveMoveDataFn.execute(gmax.pokemonShowdownId(), gmax.gmaxMove());
-                    }
+        ServerLifecycleEvents.SERVER_STARTED.register((server) -> Cobblemon.INSTANCE.getShowdownThread().queue(showdownService -> {
+            if (showdownService instanceof GraalShowdownService service) {
+                Value receiveMoveDataFn = service.context.getBindings("js").getMember("receiveCustomGmaxMove");
+                for (MaxGimmick gmax : MegaShowdownDatapackRegister.GMAX_REGISTRY) {
+                    receiveMoveDataFn.execute(gmax.pokemonShowdownId(), gmax.gmaxMove());
                 }
-                return Unit.INSTANCE;
-            });
-        });
+            }
+            return Unit.INSTANCE;
+        }));
     }
 }