+++ fabric/src/main/java/com/github/yajatkaul/mega_showdown/fabric/MegaShowdownFabricClient.java
@@ -2,7 +2,9 @@ package com.github.yajatkaul.mega_showdown.fabric;
 
 import com.github.yajatkaul.mega_showdown.MegaShowdown;
 import com.github.yajatkaul.mega_showdown.MegaShowdownClient;
+import com.github.yajatkaul.mega_showdown.block.MegaShowdownBlockEntities;
 import com.github.yajatkaul.mega_showdown.block.MegaShowdownBlocks;
+import com.github.yajatkaul.mega_showdown.block.block_entity.renderer.PedestalBlockEntityRenderer;
 import com.github.yajatkaul.mega_showdown.screen.MegaShowdownMenuTypes;
 import com.github.yajatkaul.mega_showdown.screen.custom.ZygardeCubeScreen;
 import net.fabricmc.api.ClientModInitializer;
@@ -12,6 +14,7 @@ import net.fabricmc.fabric.api.resource.ResourcePackActivationType;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.client.gui.screens.MenuScreens;
 import net.minecraft.client.renderer.RenderType;
+import net.minecraft.client.renderer.blockentity.BlockEntityRenderers;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 
@@ -39,5 +42,7 @@ public final class MegaShowdownFabricClient implements ClientModInitializer {
                 Component.translatable("message.mega_showdown.region_bias_msd"),
                 ResourcePackActivationType.NORMAL
         );
+
+        BlockEntityRenderers.register(MegaShowdownBlockEntities.PEDESTAL_BLOCK_ENTITY.get(), PedestalBlockEntityRenderer::new);
     }
 }